requirements.txt
Django==5.1.3
python-dotenv==1.0.1
djangorestframework==3.16.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
